Display tags on the map to deselect a foodtype
This allows the user to better understand that tags filters are
selected, which is otherwise confusing when the main menu is not open.
The tags can be pressed directly to remove them without using the main
menu.

The position of the tags changes based when the main panel is open,
such that they are always visible.