Refactor "food type" filtering such that logic is in MacCtrl
This commit centralizes and simplifies the way filters are applied to
the list of places and markers.  This allows to combine multiple
filters (food type, search field, etc.) in a way that will always affect
both the map and the list of search results.  This new structure will be
levereged in later commits.